feat: Dashboard property filter
