--- apps/client/src/routes/dashboards/dashboards-index/components/dashboard-table-header/index.ts
@@ -1,1 +1,0 @@
-export { DashboardsTableHeader } from './dashboards-table-header';
